backend/requirements.txt
# Fluxnote AI Backend — Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# File upload support
python-multipart==0.0.20

# HTTP client (for OpenRouter streaming)
httpx==0.28.1

# Database
supabase==2.10.0

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Environment variables
python-dotenv==1.0.1

# File parsing
PyMuPDF==1.25.1          # PDF
python-docx==1.1.2       # DOCX
python-pptx==1.0.2       # PPTX
openpyxl==3.1.5          # CSV/XLSX (fallback)

# Async file I/O
aiofiles==24.1.0
